first part of refactoring
